Move ResponseBuilder to response package. Change methods to be static
