+++ src/main/java/httpserver/http/route/requestmethod/GetMethodHandler.java
@@ -1,15 +1,16 @@
 package httpserver.http.route.requestmethod;
 
 import httpserver.http.StatusCode;
+import httpserver.http.response.ResponseBuilder;
 
 public class GetMethodHandler extends ResponseBuilder {
 
     public String getResponse(String requestPath) {
         switch (requestPath) {
             case "/simple_get":
-                return buildResponse(StatusCode.OK, null);
+                return ResponseBuilder.buildResponse(StatusCode.OK, null);
             default:
-                return buildNotFoundResponse();
+                return ResponseBuilder.buildNotFoundResponse();
         }
     }
 }
