+++ src/main/java/httpserver/http/route/requestmethod/OptionsMethodHandler.java
@@ -1,17 +1,20 @@
 package httpserver.http.route.requestmethod;
 
 import httpserver.http.StatusCode;
+import httpserver.http.response.ResponseBuilder;
 
 public class OptionsMethodHandler extends ResponseBuilder {
 
     public String getResponse(String requestPath) {
         switch (requestPath) {
             case "/method_options":
-                return buildResponse(StatusCode.OK, "Allow: GET, HEAD, OPTIONS");
+                return ResponseBuilder.buildResponse
+                        (StatusCode.OK, "Allow: GET, HEAD, OPTIONS");
             case "/method_options2":
-                return buildResponse(StatusCode.OK, "Allow: GET, HEAD, OPTIONS, PUT, POST");
+                return ResponseBuilder.buildResponse
+                        (StatusCode.OK, "Allow: GET, HEAD, OPTIONS, PUT, POST");
             default:
-                return buildNotFoundResponse();
+                return ResponseBuilder.buildNotFoundResponse();
         }
     }
 }
